requirements.txt
kafka-python
pymongo
pandas
matplotlib
streamlit
streamlit-autorefresh
